requirements.txt
fastapi>=0.109.0,<1
uvicorn[standard]>=0.27.0,<1
python-multipart>=0.0.6
pypdf>=4.0.0
python-docx>=1.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
openpyxl>=3.1.0
reportlab>=4.0.0

# optional dev
pytest>=7.4.0
pytest-asyncio>=0.23.0
